Fixed Bug in registration.bsh and timelapse_registration.bsh
BUG: "Single Channel" is not a valid choice for "Process_channel"
When choosing Single channel in reg_process_channel then
this is not a valid input for the macro.
Now reg_process_channel is set to All Channel for Single
Channel processing

External transformation has a bug it breaks the .xml file
External tranformation is disabled for now.

Added new cluster settings for resave_hdf5_output in cluster.json